ansible.cfg, ansible/hosts: Initial Ansible setup and inventory file
We generally expect ansible to be run from the top-level directory
of this repo and paths etc. are relative to that.

In `hosts` We create a `vagrant` group with one host, `default`. The port is
defined on a per-host basis, but the rest of the variables are defined
for all hosts in the `vagrant` group.

Note that the port may not be correct, depending on what other Vagrant
guests happen to be running. If we do much manual work we need to
figure out a way to work around this, maybe with data from the
`vagrant port --machine-readable` command. (Vagrant handles it
automatically, however, so this may not be an issue.)