Use ansible-galaxy to bring in etckeeper role
The `officel.etckeeper` role doesn't list a particular EPEL role as a
dependency because there are many different ones. So we need to list
our `geerlingguy.repo-epel` role explicitly in our playbook.

Note that, as per the comments in `ansible/requirements.yml`, Vagrant
knows how to download our requirements automatically, but if you are
are going to do manual runs before Vagrant has done a provisioning
run, you'll need to manually download the requirements. (Possibly we
can fix this by creating our own etckeeper role that serves only to
bring in the dependencies we need.)